Factors out db configuration to a new file
