+++ server/db/dbModel.js
@@ -1,27 +1,11 @@
-var USE_DEPLOYED_DB = false; // set this to true to use tr-pinpoint-server Heroku app's database
-
-var LOCAL_DB = {
-  name: 'pinpointdb',
-  username: 'postgres',
-  password: '',
-  options: { dialect: 'postgres', logging: false }
-};
-var DEPLOYED_DB = {
-  url: "postgres://xcjdlmsnudlseg:6V2Hd0n_ICoe-ZhVlYx2jCSDk5@ec2-54-225-199-245.compute-1.amazonaws.com:5432/df5pdun0nqbaat",
-  options: { dialect: 'postgres', dialectOptions: { ssl: true }, logging: false }
-};
+// Uncomment one of the following lines:
+// var DB_ENV = 'automatic'; // use whichever db the server is running from
+var DB_ENV = 'production'; // forcibly use deployed db on Heroku
 
 var Sequelize = require('sequelize');
 
 // Connect to either deployed or local database
-var sequelize = null;
-if (process.env.DATABASE_URL) {
-  sequelize = new Sequelize(process.env.DATABASE_URL);
-} else if (USE_DEPLOYED_DB) {
-  sequelize = new Sequelize(DEPLOYED_DB.url, DEPLOYED_DB.options);
-} else {
-  sequelize = new Sequelize(LOCAL_DB.name, LOCAL_DB.username, LOCAL_DB.password, LOCAL_DB.options);
-}
+var sequelize = require('./config')(Sequelize, DB_ENV);
 
 // Define schemas
 var Users = sequelize.define('Users', {
